feat: Add publish and unpublish methods to FilesController
This commit adds the `putPublish` and `putUnpublish` methods to the FilesController class. These methods allow users to publish or unpublish a file, making it public or private, respectively. The methods are implemented using the `_publishOrUnpublish` method, which takes care of updating the file's `isPublic` flag in the database. This feature enhances the functionality of the application by providing users with the ability to control the visibility of their files.