chore: Increase the status check timeout to 5 seconds
I am of the notion that this might be impacting how the checker is
running. Currently, I made my API return the status under 200ms.

This way, I am able to know when the database is down fast. But this
update increases it with the hopes of maybe having the checker to pass